Remove redundant type casts.

+++ core/src/test/java/ch/cyberduck/core/date/RFC1123DateFormatterTest.java
@@ -21,7 +21,7 @@ public class RFC1123DateFormatterTest {
 
     @Test
     public void testPrint() {
-        assertEquals("Thu, 01 Dec 1994 17:00:00 CET", new RFC1123DateFormatter().format((long) 786297600000L, TimeZone.getTimeZone("Europe/Zurich")));
-        assertEquals("Thu, 01 Dec 1994 16:00:00 UTC", new RFC1123DateFormatter().format((long) 786297600000L, TimeZone.getTimeZone("UTC")));
+        assertEquals("Thu, 01 Dec 1994 17:00:00 CET", new RFC1123DateFormatter().format(786297600000L, TimeZone.getTimeZone("Europe/Zurich")));
+        assertEquals("Thu, 01 Dec 1994 16:00:00 UTC", new RFC1123DateFormatter().format(786297600000L, TimeZone.getTimeZone("UTC")));
     }
 }
